requirements.txt
aiohttp>=3.9.5
aiogram>=3.7.0
aiolimiter>=1.1.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
